Fist batch of rewrite
